check discord id on signup as well as on callback.
this should've been done in the initial design. even though this *technically*
is not an error because the discord id could only be inserted into the encrypted
session by the server, checking it on signup has quite a few notable benefits.

this commit also cleans up some responses so that we get html/text responses
in the correct places.

the functions are still big and ugly but they're getting better.